Moved transcript section with it's deps one level deeper
