+++ plugin-hrm-form/src/components/contact/TranscriptSection/index.ts
@@ -1,0 +1,1 @@
+export { default as TranscriptSection } from './TranscriptSection';
